+++ crates/cli/src/commands/tool/info.rs
@@ -2,7 +2,9 @@ use crate::helpers::ProtoResource;
 use crate::printer::Printer;
 use clap::Args;
 use miette::IntoDiagnostic;
-use proto_core::{detect_version, ExecutableLocation, Id, PluginLocator};
+use proto_core::{
+    detect_version, ExecutableLocation, Id, PluginLocator, ProtoToolConfig, ToolManifest,
+};
 use proto_pdk_api::ToolMetadataOutput;
 use serde::Serialize;
 use starbase::system;
@@ -13,11 +15,13 @@ use std::path::PathBuf;
 #[derive(Serialize)]
 pub struct ToolInfo {
     bins: Vec<ExecutableLocation>,
+    config: ProtoToolConfig,
     exe_path: PathBuf,
     globals_dir: Option<PathBuf>,
     globals_prefix: Option<String>,
     id: Id,
     inventory_dir: PathBuf,
+    manifest: ToolManifest,
     metadata: ToolMetadataOutput,
     name: String,
     plugin: PluginLocator,
@@ -42,9 +46,13 @@ pub async fn info(args: ArgsRef<ToolInfoArgs>, proto: ResourceRef<ProtoResource>
     tool.create_executables(false, false).await?;
     tool.locate_globals_dir().await?;
 
+    let mut config = proto.env.load_config()?.to_owned();
+    let tool_config = config.tools.remove(&tool.id).unwrap_or_default();
+
     if args.json {
         let info = ToolInfo {
             bins: tool.get_bin_locations()?,
+            config: tool_config,
             exe_path: tool.get_exe_path()?.to_path_buf(),
             globals_dir: tool.get_globals_bin_dir().map(|p| p.to_path_buf()),
             globals_prefix: tool.get_globals_prefix().map(|p| p.to_owned()),
@@ -52,6 +60,7 @@ pub async fn info(args: ArgsRef<ToolInfoArgs>, proto: ResourceRef<ProtoResource>
             shims: tool.get_shim_locations()?,
             id: tool.id,
             name: tool.metadata.name.clone(),
+            manifest: tool.manifest,
             metadata: tool.metadata,
             plugin: tool.locator.unwrap(),
         };
@@ -112,8 +121,31 @@ pub async fn info(args: ArgsRef<ToolInfoArgs>, proto: ResourceRef<ProtoResource>
             Some(color::failure("None")),
         );
 
+        let mut versions = tool.manifest.installed_versions.iter().collect::<Vec<_>>();
+        versions.sort();
+
+        p.entry_list(
+            "Installed versions",
+            versions
+                .iter()
+                .map(|version| color::hash(version.to_string())),
+            Some(color::failure("None")),
+        );
+
         Ok(())
     })?;
+
+    // CONFIG
+
+    if !tool_config.config.is_empty() {
+        printer.named_section("Configuration", |p| {
+            for (key, value) in tool_config.config {
+                p.entry(key, value.to_string());
+            }
+
+            Ok(())
+        })?;
+    }
 
     // PLUGIN